Adding Cypress tests, rename pages, mock tests
