+++ cypress.json
@@ -1,1 +1,4 @@
-{}
+{
+  "baseUrl": "http://localhost:8000",
+  "video": false
+}
