+++ cypress/fixtures/selectors.json
@@ -1,1 +1,8 @@
-{}
+{
+  "subtitle": "h2",
+  "paragraph": "p",
+  "weatherWidget": ".display-weather",
+  "image": ".image",
+  "imageParagraph": ".image p",
+  "foother": "footer"
+}
