+++ cypress/integration/home.spec.js
@@ -5,8 +5,8 @@ const viewport = ['macbook-15', 'iphone-6', 'ipad-2'];
 describe('DownTheCode Home Page', () => {
   before(() => {
     cy.visit('http://localhost:8000/');
-    cy.title().should('equal', '');
-    cy.get('h1').should('contain', '');
+    cy.title().should('equal', 'Down the Code');
+    cy.get('h1').should('contain', 'Down the Code');
   });
 
   viewport.forEach(view => {
@@ -14,83 +14,90 @@ describe('DownTheCode Home Page', () => {
       cy.viewport(view);
     });
 
-    it(`Should display the home page title in the view: ${view}`, () => {
-      // cy
-      //   .get(selector.toggleMenuSingle)
-      //   .click()
-      //   .click()
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible')
-      //   // to be sure it says focus and the menu is closed
-      //   .get(selector.singleSelectSingleInput)
-      //   .trigger('mousedown')
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible');
-    });
-    it(`Should display the correct navigation menu in the view: ${view}`, () => {
-      // cy
-      //   .get(selector.toggleMenuSingle)
-      //   .click()
-      //   .click()
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible')
-      //   // to be sure it says focus and the menu is closed
-      //   .get(selector.singleSelectSingleInput)
-      //   .trigger('mousedown')
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible');
-    });
-    it(`Should display the side bar elements in the view: ${view}`, () => {
-      // cy
-      //   .get(selector.toggleMenuSingle)
-      //   .click()
-      //   .click()
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible')
-      //   // to be sure it says focus and the menu is closed
-      //   .get(selector.singleSelectSingleInput)
-      //   .trigger('mousedown')
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible');
-    });
-    it(`Should display the weather for the location XX in the view: ${view}`, () => {
-      // cy
-      //   .get(selector.toggleMenuSingle)
-      //   .click()
-      //   .click()
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible')
-      //   // to be sure it says focus and the menu is closed
-      //   .get(selector.singleSelectSingleInput)
-      //   .trigger('mousedown')
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible');
-    });
-    it(`Should display the Coogee image in the view: ${view}`, () => {
-      // cy
-      //   .get(selector.toggleMenuSingle)
-      //   .click()
-      //   .click()
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible')
-      //   // to be sure it says focus and the menu is closed
-      //   .get(selector.singleSelectSingleInput)
-      //   .trigger('mousedown')
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible');
-    });
-    it(`Should display the footer in the view: ${view}`, () => {
-      // cy
-      //   .get(selector.toggleMenuSingle)
-      //   .click()
-      //   .click()
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible')
-      //   // to be sure it says focus and the menu is closed
-      //   .get(selector.singleSelectSingleInput)
-      //   .trigger('mousedown')
-      //   .get(selector.menuSingle)
-      //   .should('not.be.visible');
+    it(`Should display the home page in the view: ${view}`, () => {
+      cy.get(selector.subtitle)
+        .should('contain', 'Welcome to Down the Code!')
+        .get(selector.paragraph)
+        .should(
+          'contain',
+          'You may find couple of things on this website from blogs to pictures or data projects. Enjoy!',
+        )
+        // TODO: https://github.com/RajaBellebon/tulec/issues/12
+        // .get(selector.weatherWidget)
+        // .should('.be.visible')
+        // .get(selector.image)
+        // .should('.exist')
+        .get(selector.imageParagraph)
+        .should('contain', 'Coogee Beach at Sunrise');
+      // .get(selector.foother)
+      // .should('.exist')
     });
   });
 });
+// TODO: https://github.com/RajaBellebon/tulec/issues/11
+//   it(`Should display the correct navigation menu in the view: ${view}`, () => {
+//     // cy
+//     //   .get(selector.toggleMenuSingle)
+//     //   .click()
+//     //   .click()
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible')
+//     //   // to be sure it says focus and the menu is closed
+//     //   .get(selector.singleSelectSingleInput)
+//     //   .trigger('mousedown')
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible');
+//   });
+//   it(`Should display the side bar elements in the view: ${view}`, () => {
+//     // cy
+//     //   .get(selector.toggleMenuSingle)
+//     //   .click()
+//     //   .click()
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible')
+//     //   // to be sure it says focus and the menu is closed
+//     //   .get(selector.singleSelectSingleInput)
+//     //   .trigger('mousedown')
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible');
+//   });
+//   it(`Should display the weather for the location XX in the view: ${view}`, () => {
+//     // cy
+//     //   .get(selector.toggleMenuSingle)
+//     //   .click()
+//     //   .click()
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible')
+//     //   // to be sure it says focus and the menu is closed
+//     //   .get(selector.singleSelectSingleInput)
+//     //   .trigger('mousedown')
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible');
+//   });
+//   it(`Should display the Coogee image in the view: ${view}`, () => {
+//     // cy
+//     //   .get(selector.toggleMenuSingle)
+//     //   .click()
+//     //   .click()
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible')
+//     //   // to be sure it says focus and the menu is closed
+//     //   .get(selector.singleSelectSingleInput)
+//     //   .trigger('mousedown')
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible');
+//   });
+//   it(`Should display the footer in the view: ${view}`, () => {
+//     // cy
+//     //   .get(selector.toggleMenuSingle)
+//     //   .click()
+//     //   .click()
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible')
+//     //   // to be sure it says focus and the menu is closed
+//     //   .get(selector.singleSelectSingleInput)
+//     //   .trigger('mousedown')
+//     //   .get(selector.menuSingle)
+//     //   .should('not.be.visible');
+//   });
+// });
