+++ package.json
@@ -37,11 +37,12 @@
   "license": "MIT",
   "scripts": {
     "build": "gatsby build",
-    "cypress:open": "cypress open",
     "develop": "gatsby develop",
     "start": "npm run develop",
     "format": "prettier --write \"src/**/*.js\"",
-    "test": "jest"
+    "test": "jest",
+    "test:e2e": "cypress run",
+    "test:e2e:watch": "cypress open"
   },
   "devDependencies": {
     "@babel/core": "^7.2.2",
@@ -54,6 +55,7 @@
     "enzyme-to-json": "^3.3.5",
     "husky": "^1.2.1",
     "jest-in-case": "^1.0.2",
+    "jest-transform-css": "^2.0.0",
     "lint-staged": "^8.1.0",
     "prettier": "^1.15.2",
     "react-media": "^1.9.2"
