+++ src/pages/blogs.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Layout from '../components/layout';
+
+const Blogs = () => (
+  <Layout>
+    <h2>Blogs</h2>
+    <h3>🏗🏗This page is currently under construction.🏗🏗</h3>
+    <ol>
+      <li>Head to http://downthecode.com/</li>
+    </ol>
+  </Layout>
+);
+
+export default Blogs;
