+++ src/pages/data-projects.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Layout from '../components/layout';
+import gif from '../../awesome.gif';
+
+const DataProjects = () => (
+  <Layout>
+    <h2>Data projects</h2>
+    <h3>🏗🏗This page is currently under construction.🏗🏗</h3>
+    <img src={gif} alt="awesome" />
+  </Layout>
+);
+
+export default DataProjects;
