--- src/pages/page-2.js
@@ -1,40 +1,0 @@
-import React from 'react';
-import { graphql } from 'gatsby';
-import Img from 'gatsby-image';
-
-import Layout from '../components/layout';
-
-const Nepal = ({ data }) => (
-  <Layout>
-    <h2>Nepal: Everest Base Camp</h2>
-    <p>Let's put some nice pics from our recent travel in Nepal</p>
-    <div style={{ maxWidth: '800px', marginBottom: '1.45rem' }}>
-      {data.allFile.edges.map(img => {
-        return (
-          <Img fluid={img.node.childImageSharp.fluid} key={img.node.name} />
-        );
-      })}
-    </div>
-  </Layout>
-);
-
-export default Nepal;
-
-export const query = graphql`
-  query NepalQuery {
-    # the filter is usefull if you have multiple source-filesystem instances
-    # the name "images" is set in the gatsby-config
-    allFile(filter: { sourceInstanceName: { eq: "nepal" } }) {
-      edges {
-        node {
-          childImageSharp {
-            fluid(maxWidth: 800) {
-              ...GatsbyImageSharpFluid
-            }
-          }
-          name
-        }
-      }
-    }
-  }
-`;
