--- src/pages/page-6.js
@@ -1,40 +1,0 @@
-import React from 'react';
-import { graphql } from 'gatsby';
-import Img from 'gatsby-image';
-
-import Layout from '../components/layout';
-
-const NT = ({ data }) => (
-  <Layout>
-    <h2>Northern Territory: Larapinta trek</h2>
-    <p>Trekking, Hills, Desert</p>
-    <div style={{ maxWidth: '800px', marginBottom: '1.45rem' }}>
-      {data.allFile.edges.map(img => {
-        return (
-          <Img fluid={img.node.childImageSharp.fluid} key={img.node.name} />
-        );
-      })}
-    </div>
-  </Layout>
-);
-
-export default NT;
-
-export const query = graphql`
-  query NtQuery {
-    # the filter is usefull if you have multiple source-filesystem instances
-    # the name "images" is set in the gatsby-config
-    allFile(filter: { sourceInstanceName: { eq: "nt" } }) {
-      edges {
-        node {
-          childImageSharp {
-            fluid(maxWidth: 800) {
-              ...GatsbyImageSharpFluid
-            }
-          }
-          name
-        }
-      }
-    }
-  }
-`;
