--- src/pages/page-7.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import Layout from '../components/layout';
-
-const DataProjects = () => (
-  <Layout>
-    <h2>Data projects</h2>
-    <h3>🏗🏗🏗This page is currently under construction.🏗🏗🏗</h3>
-  </Layout>
-);
-
-export default DataProjects;
