+++ src/pages/wedding-shooting.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import { graphql } from 'gatsby';
+import Img from 'gatsby-image';
+
+import Layout from '../components/layout';
+
+const Wedding = ({ data }) => (
+  <Layout>
+    <h2>Wedding costume shooting pics</h2>
+    <p>Some nice pics in Coogee!</p>
+    <div style={{ maxWidth: '800px', marginBottom: '1.45rem' }}>
+      {data.allFile.edges.map(img => {
+        return (
+          <Img fluid={img.node.childImageSharp.fluid} key={img.node.name} />
+        );
+      })}
+    </div>
+  </Layout>
+);
+
+export default Wedding;
+
+export const query = graphql`
+  query WeddingQuery {
+    # the filter is usefull if you have multiple source-filesystem instances
+    # the name "images" is set in the gatsby-config
+    allFile(filter: { sourceInstanceName: { eq: "wedding" } }) {
+      edges {
+        node {
+          childImageSharp {
+            fluid(maxWidth: 800) {
+              ...GatsbyImageSharpFluid
+            }
+          }
+          name
+        }
+      }
+    }
+  }
+`;
